feat: v2.51.0 — reportSummary/dlqSummary methods, EventExporterService, expanded Plausible+PostHog mappings
- Add AnalyticsManager::reportSummary() for quick analytics activity overview
- Add AnalyticsManager::dlqSummary() for dead letter queue monitoring
- Add EventExporterService: catalog export (JSON, CSV), provider mappings, per-category export
- Add GET /api/analytics/export/mappings and /export/catalog.csv endpoints
- Expand Plausible event map: SaaS, ecommerce, engagement, cohort, content events
- Expand PostHog event map: account lifecycle, B2B/team, engagement reserved events
- Fix config cookie_ttl type safety (cast to int)
- Bump version to 2.51.0 across all files
- Add 18 comprehensive tests for v2.51 features

+++ config/zeroboiler.php
@@ -125,7 +125,7 @@ return [
         */
         'identity' => [
             'cookie_name' => env('ANALYTICS_IDENTITY_COOKIE', 'zb_analytics_id'),
-            'cookie_ttl' => env('ANALYTICS_IDENTITY_COOKIE_TTL', 525600), // 365 days (minutes)
+            'cookie_ttl' => (int) env('ANALYTICS_IDENTITY_COOKIE_TTL', 525600), // 365 days (minutes),
             'cookie_secure' => env('ANALYTICS_IDENTITY_COOKIE_SECURE', true),
             'cookie_samesite' => env('ANALYTICS_IDENTITY_COOKIE_SAMESITE', 'Lax'),
         ],
